template for writing to/ read responses from firebase

+++ Assets/Scripts/UI/Detection.cs
@@ -65,7 +65,7 @@ public class Detection : MonoBehaviour
             Debug.Log("B pressed. response time = " + _stopwatch.ElapsedMilliseconds);
             _text.text ="Listen...\n\n\n\n  \"B\" press detected";
             _response.response = ResponseValue.yes;
-            _response.responseTime = _stopwatch.ElapsedMilliseconds;
+            _response.responseTime = (int) _stopwatch.ElapsedMilliseconds;
             _bPressed = true;
             _stopwatch.Reset();
         }
